restructured the folders into services
