+++ RD2/gql_server/src/datasources.js
@@ -1,5 +1,5 @@
-const post = require('./datasource/posts/post')
-const user = require('./datasource/users/user')
+const post = require('./services/post/datasources/post')
+const user = require('./services/user/datasources/user')
 
 module.exports = {
   post,
