+++ RD2/gql_server/src/resolvers.js
@@ -1,10 +1,10 @@
 // Import mutations
-const postMutations = require('./resolvers/mutations/postMutations')
-const userMutations = require('./resolvers/mutations/userMutations')
+const postMutations = require('./services/post/resolvers/mutations/postMutations')
+const userMutations = require('./services/user/resolvers/mutations/userMutations')
 
 // Import queries
-const postQueries = require('./resolvers/queries/postQueries')
-const userQueries = require('./resolvers/queries/userQueries')
+const postQueries = require('./services/post/resolvers/queries/postQueries')
+const userQueries = require('./services/user/resolvers/queries/userQueries')
 
 module.exports = {
   Query: {
